Rename package directories to use company.appointments
