+++ src/itest/java/uk/gov/companieshouse/company/appointments/Runner.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse;
+package uk.gov.companieshouse.company.appointments;
 
 import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
