+++ src/main/java/uk/gov/companieshouse/company/appointments/Config.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse;
+package uk.gov.companieshouse.company.appointments;
 
 import java.util.Map;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
