+++ src/main/java/uk/gov/companieshouse/company/appointments/MessageLoggingAspect.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse;
+package uk.gov.companieshouse.company.appointments;
 
 import java.util.Map;
 import java.util.Optional;
