+++ src/main/java/uk/gov/companieshouse/company/appointments/NullService.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse;
+package uk.gov.companieshouse.company.appointments;
 
 import org.springframework.stereotype.Component;
