+++ src/main/java/uk/gov/companieshouse/company/appointments/Service.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse;
+package uk.gov.companieshouse.company.appointments;
 
 /**
  * Processes an incoming message.
