+++ src/test/java/uk/gov/companieshouse/company/appointments/TestConfig.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse;
+package uk.gov.companieshouse.company.appointments;
 
 import java.util.Map;
 import java.util.UUID;
